backend/pyproject.toml
[project]
name = "news-from-historical-people-backend"
version = "0.1.0"
description = "Django backend for News from Historical People"
requires-python = ">=3.11"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.15,<4.0",
    "django-cors-headers>=4.3,<5.0",
    "python-dotenv>=1.0,<2.0",
    "requests>=2.31,<3.0",
    "langchain-google-genai>=2.0,<3.0",
    "langchain-core>=0.3,<1.0",
    "gunicorn>=21.0,<23.0",
    "whitenoise>=6.6,<7.0",
]

[tool.uv]
package = false
